bugs fixed, code linted
