+++ app.js
@@ -1,58 +1,42 @@
-var chalk = require('chalk');
-var sequence = require('when/sequence');
-
-var createClient = require('./libs/utils/create-client');
+var Bluebird = require('bluebird');
 var config = require('./config');
-var log = require('./libs/utils/log');
+var util = require('./lib/util')
+var log = require('./lib/util/log');
 
-var moduleExport;
+util.validateConfig(config)
+config = util.buildAppConfig(config)
 
-createClient(config, function (client) {
-  global.client = client;
-  global.config = config;
+var types = config.modules.types;
 
-  var modulesList = [
-    'assets',
-    'locales',
-    'environments',
-    'content_types',
-    'entries'
-  ];
-  var _export = [];
-
-  if (process.argv.length === 3) {
-    var val = process.argv[2];
-    if (val && modulesList.indexOf(val) > -1) {
-      moduleExport = require('./libs/export/' + val + '.js');
-      _export.push(function () {
-        return moduleExport.start();
-      });
-    } else {
-      log.error('Please provide valid module name.');
-      return 0;
-    }
-  } else if (process.argv.length === 2) {
-    for (var i = 0, total = modulesList.length; i < total; i++) {
-      moduleExport = require('./libs/export/' + modulesList[i] + '.js');
-      _export.push(
-        (function (moduleExport) {
-          return function () {
-            return moduleExport.start();
-          };
-        })(moduleExport)
-      );
-    }
+if (process.argv.length === 3) {
+  var val = process.argv[2];
+  if (val && modulesList.indexOf(val) > -1) {
+    var exportedModule = require('./lib/export/' + val);
+    return exportedModule.start().then(function () {
+      log.success(val + ' was exported successfully!');
+      return;
+    }).catch(function (error) {
+      log.error('Failed to migrate ' + val);
+      log.error(error);
+    })
   } else {
-    log.error('Only one module can be exported at a time.');
+    log.error('Please provide valid module name.');
     return 0;
   }
-
-  var taskResults = sequence(_export);
-  taskResults.then(function () {
-    log.success(chalk.blue('Stack with api key: ' + config.source_stack +
-      ' has been exported succesfully!'));
-  }).catch(function () {
-    log.error(chalk.red('Failed to migrate stack: ' + config.source_stack +
-      ' please check error logs for more info'));
+} else if (process.argv.length === 2) {
+  return Bluebird.map(types, function (type) {
+    log.info('Exporting: ' + type)
+    var exportedModule = require('./lib/export/' + val);
+    return exportedModule.start()
+  }, {
+    concurrency: 1
+  }).then(function () {
+    log.success('Stack: ' + config.source_stack + ' has been exported succesfully!');
+  }).catch(function (error) {
+    log.error('Failed to migrate stack: ' + config.source_stack + ' please check error logs for more info');
+    log.error(error);
   });
-});
+} else {
+  log.error('Only one module can be exported at a time.');
+  return 0;
+}
