--- beautify.language
@@ -1,13 +1,0 @@
-{
-  "beautify.language": {
-    "js": {
-      "type": ["javascript", "json"],
-      "filename": [".jsbeautifyrc"]
-      // "ext": ["js", "json"]
-      // ^^ to set extensions to be beautified using the javascript beautifier
-    },
-    "css": ["css", "scss"],
-    "html": ["htm", "html"]
-    // ^^ providing just an array sets the VS Code file type
-  }
-}
