+++ lib/export/assets.js
@@ -1,26 +1,30 @@
-/**
- * External module Dependencies.
+/*!
+ * Contentstack Export
+ * Copyright (c) 2019 Contentstack LLC
+ * MIT Licensed
  */
+
 var nativeRequest = require('request');
 var mkdirp = require('mkdirp');
 var path = require('path');
 var fs = require('fs');
 var Promise = require('bluebird');
 var _ = require('lodash');
-/**
- * Internal module Dependencies.
- */
-var request = require('../utils/request');
-var config = require('../../config');
-var helper = require('../utils/helper');
-var log = require('../utils/log');
 
+var request = require('../util/request');
+var app = require('../../app')
+var helper = require('../util/helper');
+var log = require('../util/log');
+
+var config = app.getConfig();
 var assetConfig = config.modules.assets;
 var invalidKeys = assetConfig.invalidKeys;
+
 // The no. of assets fetched and processed in a batch
 var bLimit = assetConfig.batchLimit || 15;
+
 // The no. of asset files downloaded at a time
-var vLimit = assetConfig.downloadLimit || 10;
+var vLimit = assetConfig.downloadLimit || 3;
 var assetsFolderPath = path.resolve(config.data, assetConfig.dirName);
 var assetContentsFile = path.resolve(assetsFolderPath, 'assets.json');
 var folderJSONPath = path.resolve(assetsFolderPath, 'folders.json');
@@ -28,11 +32,11 @@ var folderJSONPath = path.resolve(assetsFolderPath, 'folders.json');
 // Create asset folder
 mkdirp.sync(assetsFolderPath);
 
-function exportAssets() {
+function exportAssets () {
   this.assetContents = {};
   this.folderData = [];
   this.requestOption = {
-    uri: client.endPoint + config.apis.assets,
+    uri: config.cdn + config.apis.assets,
     qs: {
       skip: 0,
       limit: assetConfig.limit,
@@ -43,10 +47,7 @@ function exportAssets() {
         BASE: invalidKeys
       }
     },
-    headers: {
-      api_key: config.source_stack,
-      authtoken: client.authtoken
-    },
+    headers: config.headers,
     json: true
   };
 }
@@ -67,13 +68,14 @@ exportAssets.prototype = {
           return Promise.map(assetBatches, function (batch) {
             return self.getAssetJSON(batch).then(function (assetsJSON) {
               return Promise.map(assetsJSON, function (assetJSON) {
-                return self.getVersionedAssetJSON(assetJSON.uid, assetJSON._version).then(function () {
-                  self.assetContents[assetJSON.uid] = assetJSON;
-                  log.success(
-                    'The following asset has been downloaded successfully: ' +
-                    assetJSON.uid);
-                  return;
-                }).catch(function (error) {
+                return self.getVersionedAssetJSON(assetJSON.uid, assetJSON._version).then(
+                  function () {
+                    self.assetContents[assetJSON.uid] = assetJSON;
+                    log.success(
+                      'The following asset has been downloaded successfully: ' +
+                      assetJSON.uid);
+                    return;
+                  }).catch(function (error) {
                   log.error('The following asset failed to download\n' + JSON.stringify(
                     assetJSON));
                   log.error(error);
@@ -266,7 +268,7 @@ exportAssets.prototype = {
         return resolve();
       } else {
         self.requestOptionversion = {
-          uri: client.endPoint + config.apis.assets + uid,
+          uri: config.host + config.apis.assets + uid,
           headers: self.requestOption.headers,
           qs: {
             version: version,
