+++ lib/export/content_types.js
@@ -1,20 +1,20 @@
-/**
- * External module Dependencies.
+/*!
+ * Contentstack Export
+ * Copyright (c) 2019 Contentstack LLC
+ * MIT Licensed
  */
+
 var mkdirp = require('mkdirp');
 var path = require('path');
 var chalk = require('chalk');
-
 var Promise = require('bluebird');
-//var request = Promise.promisify(require('request'));
-var request = require('../utils/request');
-/**
- * Internal module Dependencies.
- */
-var config = require('../../config');
-var helper = require('../utils/helper');
-var log = require('../utils/log');
 
+var request = require('../util/request');
+var app = require('../../app');
+var helper = require('../util/helper');
+var log = require('../util/log');
+
+var config = app.getConfig()
 var contentTypeConfig = config.modules.content_types;
 var contentTypesFolderPath = path.resolve(config.data, contentTypeConfig.dirName);
 var validKeys = contentTypeConfig.validKeys;
@@ -22,15 +22,12 @@ var validKeys = contentTypeConfig.validKeys;
 // Create folder for content types
 mkdirp.sync(contentTypesFolderPath);
 
-function ExportContentTypes() {
+function ExportContentTypes () {
   this.content_types = [];
 
   this.requestOptions = {
-    url: client.endPoint + config.apis.content_types,
-    headers: {
-      api_key: config.source_stack,
-      authtoken: client.authtoken
-    },
+    url: config.cdn + config.apis.content_types,
+    headers: config.headers,
     qs: {
       include_count: true,
       asc: 'updated_at',
@@ -47,7 +44,9 @@ ExportContentTypes.prototype = {
     return new Promise(function (resolve, reject) {
       try {
         return self.getContentTypes().then(function () {
-          return self.writeContentTypes().then(resolve).catch(reject);
+          return self.writeContentTypes()
+            .then(resolve)
+            .catch(reject);
         }).catch(reject);
       } catch (error) {
         return reject(error);
@@ -66,35 +65,28 @@ ExportContentTypes.prototype = {
     return new Promise(function (resolve, reject) {
       return request(self.requestOptions).then(function (response) {
         try {
-          if (response.statusCode === 200) {
-            if (response.body.content_types.length === 0) {
-              log.success(chalk.green('No content types were found in the Stack'));
-              return resolve();
-            }
-            response.body.content_types.forEach(function (content_type) {
-              for (var key in content_type) {
-                if (validKeys.indexOf(key) === -1) {
-                  delete content_type[key];
-                }
+          if (response.body.content_types.length === 0) {
+            log.success(chalk.green('No content types were found in the Stack'));
+            return resolve();
+          }
+          response.body.content_types.forEach(function (content_type) {
+            for (var key in content_type) {
+              if (validKeys.indexOf(key) === -1) {
+                delete content_type[key];
               }
-              self.content_types.push(content_type);
-            });
-
-            skip += config.modules.content_types.limit;
-            if (skip > response.body.count) {
-              return resolve();
-            } else {
-              return self.getContentTypes(skip).then(resolve).catch(reject);
             }
-          } else if (response.statusCode >= 500) {
-            // Retry
-            return self.getContentTypes(skip).then(resolve).catch(reject);
+            self.content_types.push(content_type);
+          });
+
+          skip += config.modules.content_types.limit;
+          if (skip > response.body.count) {
+            return resolve();
           } else {
-            return reject(response.error || response.body);
+            return self.getContentTypes(skip)
+              .then(resolve)
+              .catch(reject);
           }
         } catch (error) {
-          log.error(chalk.red('Content type export failed due to the following error: ' +
-            error));
           return reject(error);
         }
       }).catch(reject);
