+++ lib/export/entries.js
@@ -1,3 +1,9 @@
+/*!
+* Contentstack Export
+* Copyright (c) 2019 Contentstack LLC
+* MIT Licensed
+*/
+
 var fs = require('fs');
 var path = require('path');
 var _ = require('lodash');
@@ -5,11 +11,12 @@ var Promise = require('bluebird');
 var chalk = require('chalk');
 var mkdirp = require('mkdirp');
 
-var request = require('../utils/request');
-var config = require('../../config');
-var helper = require('../utils/helper');
-var log = require('../utils/log');
+var request = require('../util/request');
+var app = require('../../app');
+var helper = require('../util/helper');
+var log = require('../util/log');
 
+var config = app.getConfig();
 var entriesConfig = config.modules.entries;
 var entryFolderPath = path.resolve(config.data, config.modules.entries.dirName);
 
@@ -23,10 +30,7 @@ var locales;
 
 function exportEntries() {
   this.requestOptions = {
-    headers: {
-      api_key: config.source_stack,
-      authtoken: client.authtoken
-    },
+    headers: config.headers,
     qs: {
       include_count: true,
       limit: limit
@@ -65,10 +69,7 @@ exportEntries.prototype.start = function () {
         }).then(function () {
           log.success(chalk.blue('Entry migration completed successfully'));
           return resolve();
-        }).catch(function (error) {
-          log.error(chalk.red('Entry migration failed due to the following reason: ' + error));
-          return reject();
-        });
+        }).catch(reject);
       } else {
         log.success(chalk.blue('No content_types were found in the Stack'));
         return resolve();
@@ -83,7 +84,7 @@ exportEntries.prototype.getEntry = function (apiDetails) {
   var self = this;
   return new Promise(function (resolve, reject) {
     var requestObject = {
-      url: client.endPoint + config.apis.content_types + apiDetails.content_type + config.apis.entries +
+      url: config.cdn + config.apis.content_types + apiDetails.content_type + config.apis.entries +
         apiDetails.uid,
       method: 'GET',
       headers: self.requestOptions.headers,
@@ -97,36 +98,22 @@ exportEntries.prototype.getEntry = function (apiDetails) {
       json: true
     };
     return request(requestObject).then(function (response) {
-      if (response.statusCode === 200) {
-        var entryPath = path.join(entryFolderPath, apiDetails.locale, apiDetails.content_type,
-          response.body.entry.uid);
-        mkdirp.sync(entryPath);
-        helper.writeFile(path.join(entryPath, 'version-' + response.body.entry._version +
-          '.json'), response.body.entry);
-        log.success('Completed version backup of entry: ' + response.body.entry.uid +
-          ', version: ' + response.body.entry._version + ', content type: ' + apiDetails.content_type
-        );
-        if (--apiDetails.version !== 0) {
-          return self.getEntry(apiDetails).then(function () {
-            return resolve();
-          }).catch(function (error) {
-            return reject(error);
-          });
-        } else {
-          return resolve();
-        }
-      } else if (response.statusCode >= 500) {
-        return self.getEntry(apiDetails).then(resolve).catch(reject);
-      } else if (response.statusCode >= 400 && response.statusCode <= 499) {
-        log.error(chalk.red('Entry not found. API request made: ' + JSON.stringify(
-          requestObject), ' - API Details: ' + JSON.stringify(apiDetails)));
-        log.error(chalk.black.bgWhite(JSON.stringify(response.error) || JSON.stringify(
-          response.body)));
-        return reject(response.error || response.body);
+      var entryPath = path.join(entryFolderPath, apiDetails.locale, apiDetails.content_type,
+        response.body.entry.uid);
+      mkdirp.sync(entryPath);
+      helper.writeFile(path.join(entryPath, 'version-' + response.body.entry._version +
+        '.json'), response.body.entry);
+      log.success('Completed version backup of entry: ' + response.body.entry.uid +
+        ', version: ' + response.body.entry._version + ', content type: ' + apiDetails.content_type
+      );
+      if (--apiDetails.version !== 0) {
+        return self.getEntry(apiDetails)
+          .then(resolve)
+          .catch(reject);
+      } else {
+        return resolve();
       }
-    }).catch(function (error) {
-      return reject(error);
-    });
+    }).catch(reject);
   });
 };
 
@@ -134,7 +121,7 @@ exportEntries.prototype.getEntries = function (apiDetails) {
   var self = this;
   return new Promise(function (resolve, reject) {
     var requestObject = _.cloneDeep(self.requestOptions);
-    requestObject.uri = client.endPoint + config.apis.content_types + apiDetails.content_type + config.apis.entries;
+    requestObject.uri = config.cdn + config.apis.content_types + apiDetails.content_type + config.apis.entries;
     if (typeof apiDetails.skip !== 'number') {
       apiDetails.skip = 0;
     }
@@ -147,7 +134,6 @@ exportEntries.prototype.getEntries = function (apiDetails) {
     };
 
     return request(requestObject).then(function (response) {
-      if (response.statusCode === 200) {
         // /entries/content_type_uid/locale.json
         if (!fs.existsSync(path.join(entryFolderPath, apiDetails.content_type))) {
           mkdirp.sync(path.join(entryFolderPath, apiDetails.content_type));
@@ -179,13 +165,10 @@ exportEntries.prototype.getEntries = function (apiDetails) {
             };
             return self.getEntry(entryDetails).then(function () {
               return;
-            }).catch(function (error) {
-              throw error;
-            });
+            }).catch(reject);
           }, {
             concurrency: 1
           }).then(function () {
-
             if (apiDetails.skip > response.body.count) {
               log.success(chalk.green('Completed fetching ' + apiDetails.content_type +
                 ' content type\'s entries in ' + apiDetails.locale + ' locale'));
@@ -206,33 +189,12 @@ exportEntries.prototype.getEntries = function (apiDetails) {
             return resolve();
           } else {
             apiDetails.skip += limit;
-            return self.getEntries(apiDetails).then(function () {
-              return resolve();
-            }).catch(function (error) {
-              return reject(error);
-            });
+            return self.getEntries(apiDetails)
+              .then(resolve)
+              .catch(reject);
           }
         }
-      } else if (response.statusCode >= 500) {
-        return self.getEntries(apiDetails).then(function () {
-          return resolve();
-        }).catch(function (error) {
-          return reject(error);
-        });
-      } else if (response.statusCode >= 400 && response.statusCode <= 499) {
-        log.error(chalk.red('Content not found'));
-        log.error(chalk.black.bgWhite(JSON.stringify(response.body)));
-        return resolve();
-      } else {
-        log.error(chalk.red('Something went wrong while exporting entries'));
-        log.error(chalk.black.bgWhite(response.error || JSON.stringify(response.body)));
-        return reject(response.error || response.body);
-      }
-    }).catch(function (error) {
-      log.error(chalk.red('Unknown error while exporting entries'));
-      log.error(chalk.black.bgWhite(error));
-      return reject(error);
-    });
+    }).catch(reject);
   });
 };
