+++ lib/util/helper.js
@@ -1,6 +1,9 @@
-/**
- * External module Dependencies.
- */
+/*!
+* Contentstack Export
+* Copyright (c) 2019 Contentstack LLC
+* MIT Licensed
+*/
+
 var fs = require('fs');
 var path = require('path');
 var mkdirp = require('mkdirp');
