+++ lib/util/index.js
@@ -1,0 +1,31 @@
+/*!
+* Contentstack Export
+* Copyright (c) 2019 Contentstack LLC
+* MIT Licensed
+*/
+
+var _ = require('lodash')
+var pkg = require('../../package')
+var defaultConfig = require('../../config/default')
+
+exports.validateConfig = function(config) {
+  if (!config.host || !config.cdn) {
+    throw new Error('Host/CDN end point is missing from config');
+  }
+  if (!config.port) {
+    throw new Error('Port value is missing from config');
+  }
+  if (!config.access_token) {
+    throw new Error('Kindly provide access_token');
+  }
+};
+
+exports.buildAppConfig = function (config) {
+  config = _.merge(defaultConfig, config)
+  config.headers = {
+    api_key: config.api_key,
+    access_token: config.access_token,
+    'X-User-Agent': 'contentstack-import/v' + pkg.version
+  };
+  return config;
+}
