+++ lib/util/log.js
@@ -1,3 +1,9 @@
+/*!
+* Contentstack Export
+* Copyright (c) 2019 Contentstack LLC
+* MIT Licensed
+*/
+
 var winston = require('winston');
 var path = require('path');
 var mkdirp = require('mkdirp');
