+++ lib/util/request.js
@@ -1,24 +1,24 @@
 /*!
- * contentstack-express
- * copyright (c) Contentstack
- * MIT Licensed
- */
+* Contentstack Export
+* Copyright (c) 2019 Contentstack LLC
+* MIT Licensed
+*/
 
 'use strict';
-/*!
- * Module dependencies
- */
 
 var Bluebird = require('bluebird');
 var request = Bluebird.promisify(require('request'));
 var debug = require('debug')('util:requests');
 var pkg = require('../../package');
-var MAX_RETRY_LIMIT = 10;
+var index = require('../app');
+
+var MAX_RETRY_LIMIT = 8;
+var config = index.getConfig();
 
 config.headers = {
   api_key: config.api_key,
-  authtoken: client.authtoken,
-  'X-User-Agent': 'contentstack-import/' + pkg.version
+  access_token: config.access_token,
+  'X-User-Agent': 'contentstack-import/v' + pkg.version
 };
 
 function validate(req) {
@@ -42,9 +42,7 @@ function validate(req) {
 }
 
 var makeCall = module.exports = function(req, RETRY) {
-  // console.log("inside the request", typeof RETRY)
   return new Bluebird(function (resolve, reject) {
-  var self = this;
     try {
       validate(req);
       if (typeof RETRY !== 'number') {
